+++ vignettes/matsindf.R
@@ -16,9 +16,9 @@ head(UKEnergy2000, 2)
 ## -----------------------------------------------------------------------------
 UKEnergy2000_with_metadata <- UKEnergy2000 %>% 
   # Add a column indicating the matrix in which this entry belongs (U, V, or Y).
-  matsindf:::add_UKEnergy2000_matnames(.) %>% 
+  matsindf:::add_UKEnergy2000_matnames() %>% 
   # Add columns for row names, column names, row types, and column types.
-  matsindf:::add_UKEnergy2000_row_col_meta(.) %>% 
+  matsindf:::add_UKEnergy2000_row_col_meta() %>% 
   mutate(
     # Eliminate columns we no longer need
     Ledger.side = NULL,